# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: manageRTC HRM API
  description: |
    Human Resource Management API for the manageRTC application.

    Features include:
    - Timesheet management with approval workflow
    - Leave request management
    - Overtime request handling
    - Employee lifecycle management (promotions, resignations, terminations)
    - Attendance tracking
    - Department and designation management

    **Authentication**: All endpoints require Bearer token authentication via Clerk JWT.

    **Multi-Tenant**: The API uses company-specific database collections. Include `X-Company-Id` header or use JWT claim.
  version: 1.0.0
  contact:
    name: manageRTC Support
    email: support@manageRTC.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.managertc.com/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /promotions:
    post:
      tags: [Promotions]
      summary: Create promotion request
      description: |
        Create a new employee promotion request.

        **Role Requirements**: Admin, HR.
        **Stores**: Current position (promotionFrom) is stored when creating.
        **Updates**: Employee salary when promotion is applied.
      operationId: createPromotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeId
                - fromDesignation
                - toDesignation
                - promotionDate
              properties:
                employeeId:
                  type: string
                  format: uuid
                fromDesignation:
                  type: string
                  description: Current designation (stored at creation)
                  example: 'Senior Developer'
                toDesignation:
                  type: string
                  description: New designation
                  example: 'Team Lead'
                promotionDate:
                  type: string
                  format: date
                salary:
                  type: number
                  description: New salary after promotion
                  example: 85000
                remarks:
                  type: string
      responses:
        '201':
          description: Promotion created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from Clerk authentication.

        Include in Authorization header:
        ```
        Authorization: Bearer <token>
        ```
...
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Validation failed'
        message:
          type: string
          example: 'Entries are required'
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: 'Validation Error'
            message: 'weekStart is required'
...